+++ src/wifi/model/rrpaa-wifi-manager.cc
@@ -157,8 +157,8 @@ RrpaaWifiManager::SetupPhy (const Ptr<WifiPhy> phy)
   m_minPower = phy->GetTxPowerStart ();
   m_maxPower = phy->GetTxPowerEnd ();
   m_nPower = m_maxPower - m_minPower + 1;
-  uint32_t nModes = phy->GetNModes ();
-  for (uint32_t i = 0; i < nModes; i++)
+  uint8_t nModes = phy->GetNModes ();
+  for (uint8_t i = 0; i < nModes; i++)
     {
       WifiMode mode = phy->GetMode (i);
       WifiTxVector txVector;
